adding SSO config for dev, stage and prod
