+++ siteManagerDashboard/dev/config.js
@@ -1,0 +1,5 @@
+
+export const firebaseConfig = {
+    apiKey: "AIzaSyCoz1UxNYE53ujAkXysPTSEA-IDU9utcNQ",
+    authDomain: "nih-nci-dceg-connect-dev.firebaseapp.com"
+};
