+++ siteManagerDashboard/stage/config.js
@@ -1,0 +1,5 @@
+
+export const firebaseConfig = {
+    apiKey: "AIzaSyDjFzceC2w6e9SPrSgx8_jDWnOpHdJbTGs",
+    authDomain: "nih-nci-dceg-connect-stg-5519.firebaseapp.com"
+};
